MGR: fix a compiler warning (size_t is unsigned)
svn path=/tags/boinc_core_release_6_1_16/; revision=15076

+++ clientgui/AdvancedFrame.cpp
@@ -1189,11 +1189,11 @@ void CAdvancedFrame::OnSelectComputer(wxCommandEvent& WXUNUSED(event)) {
             // Connect to the remote machine
             wxString sHost = dlg.m_ComputerNameCtrl->GetValue(); 
             long lPort = GUI_RPC_PORT; 
-            size_t iPos = sHost.find(_(":")); 
+            int iPos = sHost.find(_(":")); 
             if (iPos != -1) { 
-                wxString sPort = sHost.substr(iPos + 1); 
+                wxString sPort = sHost.substr((size_t)iPos + 1); 
                 if (!sPort.ToLong(&lPort)) lPort = GUI_RPC_PORT; 
-                sHost.erase(iPos); 
+                sHost.erase((size_t)iPos); 
             } 
             lRetVal = pDoc->Connect(
                 sHost,
